initial listing seeds

+++ db/migrations/20180726181653_create_listings_table.js
@@ -4,11 +4,11 @@ exports.up = function(knex, Promise) {
     table.increments('id').primary();
     table.string('address');
     table.specificType('photos', 'text[]');
-    table.bigInteger('price');
+    table.float('price');
     table.integer('landlords_id');
     table.foreign('landlords_id').references('landlords.id')
-    table.bigInteger('lng');
-    table.bigInteger('lat');
+    table.float('lng');
+    table.float('lat');
     table.integer('neighbourhoods_id');
     table.foreign('neighbourhoods_id').references('neighbourhoods.id');
   });
